+++ android_webview/support_library/boundary_interfaces/src/org/chromium/support_lib_boundary/WebSettingsBoundaryInterface.java
@@ -185,7 +185,7 @@ public interface WebSettingsBoundaryInterface {
 
     boolean getIgnoreDuplicateNavEnabled();
 
-    void setIgnoreDuplicateNavThreshold(int thresholdMs);
+    void setIgnoreDuplicateNavThreshold(long thresholdMs);
 
-    int getIgnoreDuplicateNavThreshold();
+    long getIgnoreDuplicateNavThreshold();
 }
